refactor: move ui folder to components
