+++ src/app/[soorah]/[ayah]/getAyahService.ts
@@ -1,6 +1,6 @@
 import { Db } from 'mongodb'
 
-import { DataPropsLatinized } from '@/lib/types'
+import { DataPropsLatinized } from '@/helpers/types'
 import { withMongo } from '@/utility/mongodb'
 
 import { GetSoorahServiceProps } from '../getSoorahService'
