+++ src/app/layout.tsx
@@ -8,9 +8,9 @@ import Link from 'next/link'
 import { TiSocialFacebookCircular, TiSocialInstagram } from 'react-icons/ti'
 
 import LogoSvg from '@/assets/img/logo.svg'
+import { Footer } from '@/components/Footer/Footer'
 import { Form } from '@/components/Form/Form'
 import { Sidebar } from '@/components/Sidebar/Sidebar'
-import { Footer } from '@/ui/Footer/Footer'
 import { GA_TRACKING_ID } from '@/utility/gtag'
 
 import { MainMetadata, MainViewport } from './metadata'
