+++ src/components/CityAndSize/CityAndSize.tsx
