+++ src/components/ColoredText/ColoredText.test.tsx
