+++ src/components/ColoredText/ColoredText.tsx
