+++ src/components/SoorahAyah/SoorahAyah.tsx
@@ -3,9 +3,10 @@ import type { JSX } from 'react'
 import Link from 'next/link'
 import { FaExternalLinkAlt } from 'react-icons/fa'
 
-import { DisplayData } from '@/lib/types'
-import { Sajda } from '@/ui/Sajda/Sajda'
+import { DisplayData } from '@/helpers/types'
 import { buildUrl } from '@/utility/buildUrl'
+
+import { Sajda } from '../Sajda/Sajda'
 
 export type SoorahAyahProps = {
   data: DisplayData
