--- src/ui/soorahAyahTitle/soorahAyahTitle.test.ts
@@ -1,5 +1,0 @@
-import { soorahAyahTitle } from './soorahAyahTitle'
-
-test('Render SoorahAyahTitle', () => {
-  expect(soorahAyahTitle(96, 1)).toMatchSnapshot()
-})
